+++ src/sync/impl/helpers.js
@@ -4,6 +4,7 @@ import { logError, invariant } from '../../utils/common'
 
 import type { Model, Collection, Database } from '../..'
 import { type RawRecord, type DirtyRaw, sanitizedRaw } from '../../RawRecord'
+import type { SyncLog } from '../index'
 
 // Returns raw record with naive solution to a conflict based on local `_changed` field
 // This is a per-column resolution algorithm. All columns that were changed locally win
@@ -54,10 +55,29 @@ export function prepareCreateFromRaw<T: Model>(collection: Collection<T>, dirtyR
   })
 }
 
-export function prepareUpdateFromRaw<T: Model>(record: T, updatedDirtyRaw: DirtyRaw): T {
+export function prepareUpdateFromRaw<T: Model>(
+  record: T,
+  updatedDirtyRaw: DirtyRaw,
+  log: ?SyncLog,
+): T {
+  // Note COPY for log - only if needed
+  const logConflict = log && !!record._raw._changed
+  const logLocal = logConflict ? { ...record._raw } : {}
+  const logRemote = logConflict ? { ...updatedDirtyRaw } : {}
+
   const newRaw = resolveConflict(record._raw, updatedDirtyRaw)
   return record.prepareUpdate(() => {
     replaceRaw(record, newRaw)
+
+    // log resolved conflict - if any
+    if (logConflict && log) {
+      log.resolvedConflicts = log.resolvedConflicts || []
+      log.resolvedConflicts.push({
+        local: logLocal,
+        remote: logRemote,
+        resolved: { ...record._raw },
+      })
+    }
   })
 }